+++ native/ex_ratatui/src/lib.rs
@@ -4,6 +4,7 @@ mod rendering;
 mod style;
 mod terminal;
 mod test_utils;
+mod text_input;
 mod widgets;
 
 rustler::init!("Elixir.ExRatatui.Native");